# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AuditIQ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /entries/{entryId}:
    get:
      operationId: getJournalEntry
      tags: [journal-entries]
      summary: Get single journal entry with full details
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntryDetail"
  /entries/{entryId}/override:
    post:
      operationId: overrideRiskScore
      tags: [risk-scores]
      summary: Auditor override of risk score
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverrideBody"
      responses:
        "200":
          description: Override applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScore"
  /entries/{entryId}/explanation:
    get:
      operationId: getAiExplanation
      tags: [ai-explanations]
      summary: Get AI explanation for an entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExplanation"
    post:
      operationId: generateAiExplanation
      tags: [ai-explanations]
      summary: Generate/regenerate AI explanation
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI explanation generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExplanation"

  # Dashboard Analytics
...
  /engagements/{id}/report/pdf:
    get:
      operationId: downloadPdfReport
      tags: [reports]
      summary: Download PDF audit report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /engagements/{id}/report/excel:
    get:
      operationId: downloadExcelReport
      tags: [reports]
      summary: Download Excel report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary

  # Audit Logs
...
components:
  schemas:
...
    JournalEntryDetail:
      type: object
      properties:
        id:
          type: integer
        engagementId:
          type: integer
        entryDate:
          type: string
        postedBy:
          type: string
        description:
          type: string
        debitAccount:
          type: string
        creditAccount:
          type: string
        amount:
          type: number
        postingTime:
          type: string
        referenceNumber:
          type: string
        rawData:
          type: object
        riskScore:
          $ref: "#/components/schemas/RiskScore"
        aiExplanation:
          $ref: "#/components/schemas/AiExplanation"
      required: [id, engagementId, entryDate, postedBy, description, amount]
...
    RiskScore:
      type: object
      properties:
        id:
          type: integer
        entryId:
          type: integer
        totalScore:
          type: number
        riskLevel:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        postingTimeScore:
          type: number
        amountScore:
          type: number
        userConcentrationScore:
          type: number
        keywordScore:
          type: number
        frequencyScore:
          type: number
        confidenceScore:
          type: number
        overridden:
          type: boolean
        overrideReason:
          type: string
        overriddenBy:
          type: string
        overriddenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, entryId, totalScore, riskLevel, confidenceScore, overridden]
    OverrideBody:
      type: object
      properties:
        riskLevel:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        reason:
          type: string
      required: [riskLevel, reason]

    # AI Explanations
    AiExplanation:
      type: object
      properties:
        id:
          type: integer
        entryId:
          type: integer
        explanation:
          type: string
        triggers:
          type: array
          items:
            type: string
        isaReference:
          type: string
        generatedAt:
          type: string
          format: date-time
        modelUsed:
          type: string
      required: [id, entryId, explanation, triggers, generatedAt]

    # Dashboard
...